fix(atelier): resolve declaration barrel prop types
